refactor: удалить пустую строку, удалить пустые импорты, добавить проверку на пустой список, добавить логи в CleanerScheduler

+++ src/main/java/faang/school/urlshortenerservice/util/CleanerScheduler.java
@@ -1,10 +1,10 @@
 package faang.school.urlshortenerservice.util;
 
-
 import faang.school.urlshortenerservice.entity.Url;
 import faang.school.urlshortenerservice.repository.HashRepository;
 import faang.school.urlshortenerservice.repository.UrlRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -15,6 +15,7 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
+@Slf4j
 public class CleanerScheduler {
 
     @Value("${hash-cleaner.time-ago.year}")
@@ -25,11 +26,16 @@ public class CleanerScheduler {
     @Transactional
     @Scheduled(cron = "${hash-cleaner.cron}")
     public void cleanOldUrls() {
+        log.info("Starting a cleanup of the old urls");
         List<Url> urls = urlRepository.findAllByCreatedAtBefore(LocalDateTime.now().minusYears(yearsAgo));
+        if (urls.isEmpty()) {
+            return;
+        }
         List<String> hashes = urls.stream()
                 .map(Url::getHash)
                 .toList();
         hashRepository.save(hashes);
         urlRepository.deleteAllInBatch(urls);
+        log.info("The old urls are cleaned");
     }
 }
